Abstract up length, from(_:), and to(_:) to SpanningFragment
