+++ Rhythm/MetricalDurationSpanningFragment.swift
@@ -8,30 +8,4 @@
 
 import ArithmeticTools
 
-public protocol MetricalDurationSpanningFragment: SpanningFragment, MetricalDurationSpanning {
-    associatedtype Base: Fragmentable
-    var base: Base { get }
-    var range: Range<Fraction> { get }
-}
-
-extension MetricalDurationSpanningFragment {
-
-    public var length: Fraction {
-        return range.length
-    }
-}
-
-extension MetricalDurationSpanningFragment where Fragment == Self {
-
-    func from(_ offset: Fraction) -> Self {
-        assert(offset >= self.range.lowerBound)
-        let range = offset ..< self.range.upperBound
-        return self[range]
-    }
-
-    func to(_ offset: Fraction) -> Self {
-        assert(offset <= self.range.upperBound)
-        let range = self.range.lowerBound ..< offset
-        return self[range]
-    }
-}
+public protocol MetricalDurationSpanningFragment: SpanningFragment, MetricalDurationSpanning { }
