+++ Rhythm/SpanningFragment.swift
@@ -6,4 +6,32 @@
 //
 //
 
-public protocol SpanningFragment: Spanning, Fragmentable { }
+import ArithmeticTools
+
+public protocol SpanningFragment: Spanning, Fragmentable {
+    associatedtype Base: Fragmentable
+    var base: Base { get }
+    var range: Range<Fraction> { get }
+}
+
+extension SpanningFragment {
+
+    public var length: Fraction {
+        return range.length
+    }
+}
+
+extension SpanningFragment where Fragment == Self {
+
+    func from(_ offset: Fraction) -> Self {
+        assert(offset >= self.range.lowerBound)
+        let range = offset ..< self.range.upperBound
+        return self[range]
+    }
+
+    func to(_ offset: Fraction) -> Self {
+        assert(offset <= self.range.upperBound)
+        let range = self.range.lowerBound ..< offset
+        return self[range]
+    }
+}
